rf(folder-editor): Outsourced Add and Remove Skill Logik to ded. Hook
